+++ client/homescreen/activity-panel/index.js
@@ -26,6 +26,7 @@ import {
 } from './orders/utils';
 import { getAllPanels } from './panels';
 import { getUnapprovedReviews } from './reviews/utils';
+import { getUrlParams } from '../../utils';
 
 export const ActivityPanel = () => {
 	const panelsData = useSelect( ( select ) => {
@@ -69,6 +70,13 @@ export const ActivityPanel = () => {
 		return null;
 	}
 
+	const getInitialOpenState = ( panelId ) => {
+		const { opened_panel: openedPanel } = getUrlParams(
+			window.location.search
+		);
+		return panelId === openedPanel;
+	};
+
 	return (
 		<Panel className="woocommerce-activity-panel">
 			{ panels.map( ( panelData ) => {
@@ -84,7 +92,12 @@ export const ActivityPanel = () => {
 				return collapsible ? (
 					<PanelBody
 						title={ [
-							<Text key={ title } variant="title.small">
+							<Text
+								key={ title }
+								variant="title.small"
+								size="20"
+								lineHeight="28px"
+							>
 								{ title }
 							</Text>,
 							count !== null && (
@@ -96,7 +109,7 @@ export const ActivityPanel = () => {
 						] }
 						key={ id }
 						className={ className }
-						initialOpen={ initialOpen }
+						initialOpen={ getInitialOpenState( id ) || initialOpen }
 						collapsible={ collapsible }
 						disabled={ ! collapsible }
 						onToggle={ ( isOpen ) => {
@@ -119,7 +132,13 @@ export const ActivityPanel = () => {
 								aria-expanded={ false }
 								disabled={ true }
 							>
-								<Text variant="title.small">{ title }</Text>
+								<Text
+									variant="title.small"
+									size="20"
+									lineHeight="28px"
+								>
+									{ title }
+								</Text>
 								{ count !== null && <Badge count={ count } /> }
 							</Button>
 						</h2>