+++ client/layout/store-alerts/index.js
@@ -9,7 +9,6 @@ import {
 	CardBody,
 	CardFooter,
 	CardHeader,
-	Dashicon,
 	SelectControl,
 } from '@wordpress/components';
 import classnames from 'classnames';
@@ -209,10 +208,12 @@ export class StoreAlerts extends Component {
 		return (
 			<Card className={ className } size={ null }>
 				<CardHeader isBorderless>
-					<Text variant="title.medium" as="h2">
-						{ alert.icon && (
-							<Dashicon key="icon" icon={ alert.icon } />
-						) }
+					<Text
+						variant="title.medium"
+						as="h2"
+						size="24"
+						lineHeight="32px"
+					>
 						{ alert.title }
 					</Text>
 					{ numberOfAlerts > 1 && (
@@ -225,7 +226,10 @@ export class StoreAlerts extends Component {
 									'woocommerce-admin'
 								) }
 							>
-								<Icon icon={ chevronLeft } />
+								<Icon
+									icon={ chevronLeft }
+									className="arrow-left-icon"
+								/>
 							</Button>
 							<span
 								className="woocommerce-store-alerts__pagination-label"
@@ -259,7 +263,10 @@ export class StoreAlerts extends Component {
 									'woocommerce-admin'
 								) }
 							>
-								<Icon icon={ chevronRight } />
+								<Icon
+									icon={ chevronRight }
+									className="arrow-right-icon"
+								/>
 							</Button>
 						</div>
 					) }