+++ packages/components/src/index.js
@@ -1,9 +1,4 @@
-/**
- * External dependencies
- */
-import 'react-dates/initialize';
-// The above: Turn on react-dates classes/styles, see https://github.com/airbnb/react-dates#initialize
-
+export { default as AbbreviatedCard } from './abbreviated-card';
 export { default as AdvancedFilters } from './advanced-filters';
 export { default as AnimationSlider } from './animation-slider';
 export { default as Chart } from './chart';
@@ -59,10 +54,7 @@ export { default as Tag } from './tag';
 export { default as TextControl } from './text-control';
 export { default as TextControlWithAffixes } from './text-control-with-affixes';
 export { default as Timeline } from './timeline';
-export { default as useFilters } from './higher-order/use-filters';
 export { default as ViewMoreList } from './view-more-list';
 export { default as WebPreview } from './web-preview';
 export { Badge } from './badge';
-export { default as WooRemotePayment } from './woo-remote-payment';
-export { default as WooRemotePaymentForm } from './woo-remote-payment-form';
 export { DynamicForm } from './dynamic-form';