+++ packages/data/src/plugins/actions.ts
@@ -1,7 +1,12 @@
 /**
  * External dependencies
  */
-import { apiFetch, dispatch, select } from '@wordpress/data-controls';
+import {
+	apiFetch,
+	select,
+	dispatch as depreciatedDispatch,
+} from '@wordpress/data-controls';
+import { controls } from '@wordpress/data';
 import { _n, sprintf } from '@wordpress/i18n';
 
 /**
@@ -16,6 +21,12 @@ import {
 	PluginNames,
 	SelectorKeysWithActions,
 } from './types';
+
+// Can be removed in WP 5.9, wp.data is supported in >5.7.
+const dispatch =
+	controls && controls.dispatch ? controls.dispatch : depreciatedDispatch;
+const resolveSelect =
+	controls && controls.resolveSelect ? controls.resolveSelect : select;
 
 type PluginsResponse< PluginData > = {
 	data: PluginData;
@@ -35,13 +46,13 @@ type ActivatePluginsResponse = PluginsResponse< {
 } >;
 
 function isWPError(
-	error: WPError< PluginNames > | string
+	error: WPError< PluginNames > | Error | string
 ): error is WPError< PluginNames > {
 	return ( error as WPError ).errors !== undefined;
 }
 
 export function formatErrors(
-	response: WPError< PluginNames > | string
+	response: WPError< PluginNames > | Error | string
 ): string {
 	if ( isWPError( response ) ) {
 		// Replace the slug with a plugin name if a constant exists.
@@ -59,8 +70,12 @@ export function formatErrors(
 				);
 			}
 		);
+	} else if ( typeof response === 'string' ) {
+		return response;
+	} else {
+		return response.message;
 	}
-	return response as string;
+	return '';
 }
 
 const formatErrorMessage = (
@@ -72,7 +87,7 @@ const formatErrorMessage = (
 		_n(
 			'Could not %(actionType)s %(pluginName)s plugin, %(error)s',
 			'Could not %(actionType)s the following plugins: %(pluginName)s with these Errors: %(error)s',
-			Object.keys( pluginErrors ).length,
+			Object.keys( pluginErrors ).length || 1,
 			'woocommerce-admin'
 		),
 		{
@@ -188,6 +203,10 @@ export function* installPlugins( plugins: string[] ) {
 
 		return results;
 	} catch ( error ) {
+		if ( plugins.length === 1 && ! error[ plugins[ 0 ] ] ) {
+			// Incase of a network error
+			error = { [ plugins[ 0 ] ]: error.message };
+		}
 		yield setError( 'installPlugins', error );
 		throw new Error( formatErrorMessage( error ) );
 	}
@@ -215,8 +234,12 @@ export function* activatePlugins( plugins: string[] ) {
 
 		return results;
 	} catch ( error ) {
+		if ( plugins.length === 1 && ! error[ plugins[ 0 ] ] ) {
+			// Incase of a network error
+			error = { [ plugins[ 0 ] ]: error.message };
+		}
 		yield setError( 'activatePlugins', error );
-		throw new Error( formatErrors( error ) );
+		throw new Error( formatErrorMessage( error, 'activate' ) );
 	}
 }
 
@@ -235,16 +258,20 @@ export function* installAndActivatePlugins( plugins: string[] ) {
 }
 
 export const createErrorNotice = ( errorMessage: string ) => {
-	return dispatch( 'core/notices', 'createNotice', errorMessage );
+	return dispatch( 'core/notices', 'createNotice', 'error', errorMessage );
 };
 
 export function* connectToJetpack(
 	getAdminLink: ( endpoint: string ) => string
 ) {
-	const url: string = yield select( STORE_NAME, 'getJetpackConnectUrl', {
-		redirect_url: getAdminLink( 'admin.php?page=wc-admin' ),
-	} );
-	const error: string = yield select(
+	const url: string = yield resolveSelect(
+		STORE_NAME,
+		'getJetpackConnectUrl',
+		{
+			redirect_url: getAdminLink( 'admin.php?page=wc-admin' ),
+		}
+	);
+	const error: string = yield resolveSelect(
 		STORE_NAME,
 		'getPluginsError',
 		'getJetpackConnectUrl'