+++ packages/data/CHANGELOG.md
@@ -1,3 +1,12 @@
+# Unreleased
+
+-   Fix error parsing of plugin data package. #7164
+-   Update dependencies
+
+# 1.3.1
+
+-   Fix, state md5 as npm dependency. #7087
+
 # 1.3.0
 
 -   Fix parsing bad JSON string data in useUserPreferences hook. #6819