+++ packages/dependency-extraction-webpack-plugin/CHANGELOG.md
@@ -1,3 +1,7 @@
+# 1.6.0
+
+-   Add new `bundledPackages` option to bundle in specific packages.
+
 # 1.5.0
 
 -   Add `@woocommerce/explat` to list of packages.